implemented leaderboard feature
